Refactor:Move find DtoAttribute in model to model extension method
This commit does not contain any functional change.

This commit is part of #15.
